added prms in the modeling dashboard
added prms in the modeling dashboard. also moved all the react component
to separate files.
