+++ .gitignore
@@ -13,3 +13,4 @@ app/visualization/
 .tmp
 app/static/bower_components
 node_modules
+deployment-config.json
